# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ZQAI API",
    "description": "REST API for the ZQAI portal (news, events, managers, departments, compliance officer, reception schedule, file uploads and user/auth management). Backend: Express 5 + PostgreSQL.\n\nNOTE: BearerAuth is documented as the intended security scheme for state-changing and administrative operations, but the current implementation does NOT enforce any authentication middleware. The live scan is expected to reveal this as a broken-authentication / broken-function-level-authorization finding.",
    "version": "0.0.0"
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Local development server" }
  ],
...
  "paths": {
...
    "/api/news": {
      "get": {
        "operationId": "getNews",
        "summary": "List news articles",
        "tags": ["News"],
        "parameters": [
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "format": "int64" } },
          { "name": "lang", "in": "query", "required": false, "schema": { "type": "string", "enum": ["ru", "kk", "en"] } }
        ],
        "responses": {
          "200": {
            "description": "List of news",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/News" } } } }
          },
          "500": { "description": "Failed to fetch news" }
        }
      },
      "post": {
        "operationId": "createNews",
        "summary": "Create a news article (should be authenticated)",
        "tags": ["News"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NewsCreateRequest" } } }
        },
        "responses": {
          "201": {
            "description": "News created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NewsCreated" } } }
          },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Failed to create news" }
        }
      }
    },
    "/api/news/{id}": {
      "get": {
        "operationId": "getNewsById",
        "summary": "Get a single news article with images",
        "tags": ["News"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": {
            "description": "News item",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/News" } } }
          },
          "404": { "description": "News not found" },
          "500": { "description": "Failed to fetch news detail" }
        }
      },
      "put": {
        "operationId": "updateNews",
        "summary": "Update a news article (should be authenticated)",
        "tags": ["News"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NewsUpdateRequest" } } }
        },
        "responses": {
          "200": {
            "description": "News updated",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Message" } } }
          },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "News not found" },
          "500": { "description": "Failed to update news" }
        }
      },
      "delete": {
        "operationId": "deleteNews",
        "summary": "Delete a news article and its images (should be authenticated)",
        "tags": ["News"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": {
            "description": "News deleted",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Message" } } }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "News not found" },
          "500": { "description": "Failed to delete news" }
        }
      }
    },
    "/api/news/{id}/images": {
      "post": {
        "operationId": "addNewsImages",
        "summary": "Add images to an existing news article (should be authenticated)",
        "tags": ["News"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NewsImagesRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Images added",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Message" } } }
          },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Failed to add images" }
        }
      }
    },
    "/api/news/{id}/images/{imgId}": {
      "delete": {
        "operationId": "deleteNewsImage",
        "summary": "Remove a single image from a news article (should be authenticated)",
        "tags": ["News"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } },
          { "name": "imgId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": {
            "description": "Image deleted",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Message" } } }
          },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Failed to delete image" }
        }
      }
    },
    "/api/managers": {
      "get": {
        "operationId": "getManagers",
        "summary": "List managers",
        "tags": ["Managers"],
        "parameters": [
          { "name": "lang", "in": "query", "required": false, "schema": { "type": "string", "enum": ["ru", "kk", "en"] } }
        ],
        "responses": {
          "200": {
            "description": "List of managers",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Manager" } } } }
          },
          "500": { "description": "Failed to fetch managers" }
        }
      },
      "post": {
        "operationId": "createManager",
        "summary": "Create a manager (should be authenticated)",
        "tags": ["Managers"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ManagerRequest" } } }
        },
        "responses": {
          "201": {
            "description": "Manager created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Manager" } } }
          },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Failed to create manager" }
        }
      }
    },
...
    "/api/events": {
      "get": {
        "operationId": "getEvents",
        "summary": "List events",
        "tags": ["Events"],
        "parameters": [
          { "name": "lang", "in": "query", "required": false, "schema": { "type": "string", "enum": ["ru", "kk", "en"] } }
        ],
        "responses": {
          "200": {
            "description": "List of events",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Event" } } } }
          },
          "500": { "description": "Failed to fetch events" }
        }
      },
      "post": {
        "operationId": "createEvent",
        "summary": "Create an event (should be authenticated)",
        "tags": ["Events"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EventRequest" } } }
        },
        "responses": {
          "201": {
            "description": "Event created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EventWrapper" } } }
          },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Failed to create event" }
        }
      }
    },
...
    "/api/departments": {
      "get": {
        "operationId": "getDepartments",
        "summary": "List departments with nested divisions and sectors",
        "tags": ["Departments"],
        "responses": {
          "200": {
            "description": "List of departments",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Department" } } } }
          },
          "500": { "description": "Failed to fetch departments" }
        }
      }
    },
...
    "/api/compliance-officer": {
      "get": {
        "operationId": "getComplianceOfficer",
        "summary": "Get compliance officer info",
        "tags": ["ComplianceOfficer"],
        "responses": {
          "200": {
            "description": "Compliance officer",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ComplianceOfficer" } } }
          },
          "404": { "description": "Not found" },
          "500": { "description": "Server error" }
        }
      },
      "put": {
        "operationId": "updateComplianceOfficer",
        "summary": "Create or update compliance officer info (should be authenticated)",
        "tags": ["ComplianceOfficer"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ComplianceOfficerRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Message" } } }
          },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/api/reception": {
      "get": {
        "operationId": "getReception",
        "summary": "Get reception schedule",
        "tags": ["Reception"],
        "parameters": [
          { "name": "lang", "in": "query", "required": false, "schema": { "type": "string", "enum": ["ru", "kk", "en"] } }
        ],
        "responses": {
          "200": {
            "description": "Reception schedule entries",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ReceptionEntry" } } } }
          },
          "500": { "description": "Failed to fetch reception data" }
        }
      }
    },
    "/api/logs": {
      "get": {
        "operationId": "getLogs",
        "summary": "List system audit logs (should be admin-only)",
        "tags": ["Logs"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of log entries",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/LogEntry" } } } }
          },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Failed to fetch logs" }
        }
      }
    },
    "/api/upload": {
      "post": {
        "operationId": "uploadFile",
        "summary": "Upload an image file (should be authenticated)",
        "tags": ["Upload"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": { "type": "string", "format": "binary" }
                },
                "required": ["image"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File uploaded",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UploadResponse" } } }
          },
          "400": { "description": "No file uploaded" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Server error" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT issued by POST /api/auth/login. NOTE: not currently enforced by any server middleware."
      }
    },
    "schemas": {
...
      "Message": {
        "type": "object",
        "properties": {
          "message": { "type": "string" }
        }
      },
...
      "News": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "title": { "type": "string" },
          "content": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "language": { "type": "string" },
          "image_url": { "type": "string", "nullable": true },
          "images": { "type": "array", "items": { "type": "string" } }
        }
      },
      "NewsCreateRequest": {
        "type": "object",
        "required": ["title", "content"],
        "properties": {
          "title": { "type": "string", "minLength": 1, "maxLength": 500 },
          "content": { "type": "string" },
          "images": { "type": "array", "items": { "type": "string" } },
          "userId": { "type": "integer", "format": "int64" },
          "language": { "type": "string", "enum": ["ru", "kk", "en"] }
        }
      },
      "NewsUpdateRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "maxLength": 500 },
          "content": { "type": "string" },
          "language": { "type": "string", "enum": ["ru", "kk", "en"] }
        }
      },
      "NewsCreated": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "newsId": { "type": "integer", "format": "int64" }
        }
      },
      "NewsImagesRequest": {
        "type": "object",
        "required": ["images"],
        "properties": {
          "images": { "type": "array", "items": { "type": "string" } }
        }
      },
      "Manager": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "role": { "type": "string" },
          "image_url": { "type": "string", "nullable": true },
          "photo_url": { "type": "string", "nullable": true },
          "language": { "type": "string" },
          "display_order": { "type": "integer", "format": "int64" },
          "reception_day": { "type": "string", "nullable": true },
          "reception_time": { "type": "string", "nullable": true }
        }
      },
      "ManagerRequest": {
        "type": "object",
        "required": ["name", "role"],
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 255 },
          "role": { "type": "string", "maxLength": 255 },
          "image_url": { "type": "string", "nullable": true },
          "language": { "type": "string", "enum": ["ru", "kk", "en"] },
          "display_order": { "type": "integer", "format": "int64" }
        }
      },
      "Event": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "title": { "type": "string" },
          "event_date": { "type": "string", "format": "date" },
          "time": { "type": "string", "nullable": true },
          "location": { "type": "string", "nullable": true },
          "language": { "type": "string" },
          "news_id": { "type": "integer", "format": "int64", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "EventRequest": {
        "type": "object",
        "required": ["title", "event_date"],
        "properties": {
          "title": { "type": "string", "minLength": 1, "maxLength": 255 },
          "event_date": { "type": "string", "format": "date" },
          "time": { "type": "string", "maxLength": 50, "nullable": true },
          "location": { "type": "string", "maxLength": 255, "nullable": true },
          "language": { "type": "string", "enum": ["ru", "kk", "en"] },
          "news_id": { "type": "integer", "format": "int64", "nullable": true }
        }
      },
      "EventWrapper": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "event": { "$ref": "#/components/schemas/Event" }
        }
      },
      "Sector": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "division_id": { "type": "integer", "format": "int64" },
          "name_ru": { "type": "string" },
          "name_kk": { "type": "string" },
          "head_full_name": { "type": "string", "nullable": true },
          "head_title_ru": { "type": "string", "nullable": true },
          "head_title_kk": { "type": "string", "nullable": true }
        }
      },
      "Division": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "department_id": { "type": "integer", "format": "int64" },
          "name_ru": { "type": "string" },
          "name_kk": { "type": "string" },
          "head_full_name": { "type": "string", "nullable": true },
          "head_title_ru": { "type": "string", "nullable": true },
          "head_title_kk": { "type": "string", "nullable": true },
          "sectors": { "type": "array", "items": { "$ref": "#/components/schemas/Sector" } }
        }
      },
      "Department": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name_ru": { "type": "string" },
          "name_kk": { "type": "string" },
          "director_full_name": { "type": "string", "nullable": true },
          "director_title_ru": { "type": "string", "nullable": true },
          "director_title_kk": { "type": "string", "nullable": true },
          "divisions": { "type": "array", "items": { "$ref": "#/components/schemas/Division" } }
        }
      },
      "ComplianceOfficer": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "full_name": { "type": "string" },
          "position": { "type": "string" },
          "phone": { "type": "string", "nullable": true },
          "email": { "type": "string", "format": "email", "nullable": true },
          "reception_schedule": { "type": "string", "nullable": true },
          "photo_url": { "type": "string", "nullable": true },
          "description": { "type": "string", "nullable": true },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "ComplianceOfficerRequest": {
        "type": "object",
        "required": ["full_name", "position"],
        "properties": {
          "full_name": { "type": "string", "minLength": 1, "maxLength": 255 },
          "position": { "type": "string", "maxLength": 255 },
          "phone": { "type": "string", "maxLength": 50, "nullable": true },
          "email": { "type": "string", "format": "email", "nullable": true },
          "reception_schedule": { "type": "string", "nullable": true },
          "photo_url": { "type": "string", "nullable": true },
          "description": { "type": "string", "nullable": true }
        }
      },
      "ReceptionEntry": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "role": { "type": "string" },
          "photo_url": { "type": "string", "nullable": true },
          "image_url": { "type": "string", "nullable": true },
          "reception_day": { "type": "string", "nullable": true },
          "reception_time": { "type": "string", "nullable": true },
          "display_order": { "type": "integer", "format": "int64" }
        }
      },
      "LogEntry": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "event_type": { "type": "string" },
          "user_id": { "type": "integer", "format": "int64", "nullable": true },
          "details": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "UploadResponse": {
        "type": "object",
        "properties": {
          "imageUrl": { "type": "string" }
        }
      }
    }
  }
}